api/requirements.txt
numpy
Pillow
scikit-learn
torch
torchvision
uvicorn
fastapi
pydantic
joblib

python-multipart
